// src/app/shared/model/rubroInfrastructura.ts
export class RubroInfraestructura {
    idRubro: number;
    nombrePredioObjInversion: string;
    fechaInicioEjecucion: Date;
    fechaFinEjecucion: Date;
    largo: number;
    ancho: number;
    circunferencia: number;
    profundidad: number;
    altura: number;
    calculoArea: number;
    idTipoInfraestructura: number;
    numeroDivisiones: number;
}